[test] Add group-unaligned
Test case for:
   Bug 22441 -- Unexpected shift with push_group and pop_group
   https://bugs.freedesktop.org/show_bug.cgi?id=22441

This is a test that demonstrates the error in the pdf backend when using
groups on surfaces with non-integer sizes. In order to create such a
surface, we need to update the boilerplate to use doubles instead of
integers when specifying the surface size.

+++ boilerplate/cairo-boilerplate-beos-private.h
@@ -10,10 +10,10 @@ CAIRO_BEGIN_DECLS
 extern cairo_surface_t *
 _cairo_boilerplate_beos_create_surface (const char              *name,
 					cairo_content_t          content,
-					int                      width,
-					int                      height,
-					int                      max_width,
-					int                      max_height,
+					double                      width,
+					double                      height,
+					double                      max_width,
+					double                      max_height,
 					cairo_boilerplate_mode_t mode,
 					int                      id,
 					void                   **closure);
@@ -24,10 +24,10 @@ _cairo_boilerplate_beos_cleanup (void* closure);
 extern cairo_surface_t *
 _cairo_boilerplate_beos_create_surface_for_bitmap (const char              *name,
 						   cairo_content_t          content,
-						   int                      width,
-						   int                      height,
-						   int                      max_width,
-						   int                      max_height,
+						   double                      width,
+						   double                      height,
+						   double                      max_width,
+						   double                      max_height,
 						   cairo_boilerplate_mode_t mode,
 						   int                      id,
 						   void                   **closure);
