+++ boilerplate/cairo-boilerplate-glitz-private.h
@@ -41,10 +41,10 @@ typedef struct _glitz_target_closure_base {
 cairo_surface_t *
 _cairo_boilerplate_glitz_glx_create_surface (const char			 *name,
 					     cairo_content_t		  content,
-					     int			  width,
-					     int			  height,
-					     int			  max_width,
-					     int			  max_height,
+					     double			  width,
+					     double			  height,
+					     double			  max_width,
+					     double			  max_height,
 					     cairo_boilerplate_mode_t	  mode,
 					     int                          id,
 					     void			**closure);
@@ -57,10 +57,10 @@ _cairo_boilerplate_glitz_glx_cleanup (void *closure);
 cairo_surface_t *
 _cairo_boilerplate_glitz_agl_create_surface (const char			 *name,
 					     cairo_content_t		  content,
-					     int			  width,
-					     int			  height,
-					     int			  max_width,
-					     int			  max_height,
+					     double			  width,
+					     double			  height,
+					     double			  max_width,
+					     double			  max_height,
 					     cairo_boilerplate_mode_t	  mode,
 					     int                          id,
 					     void			**closure);
@@ -73,10 +73,10 @@ _cairo_boilerplate_glitz_agl_cleanup (void *closure);
 cairo_surface_t *
 _cairo_boilerplate_glitz_wgl_create_surface (const char			 *name,
 					     cairo_content_t		  content,
-					     int			  width,
-					     int			  height,
-					     int			  max_width,
-					     int			  max_height,
+					     double			  width,
+					     double			  height,
+					     double			  max_width,
+					     double			  max_height,
 					     cairo_boilerplate_mode_t	  mode,
 					     int                          id,
 					     void			**closure);
