+++ boilerplate/cairo-boilerplate-qt.cpp
@@ -56,10 +56,10 @@ _cairo_boilerplate_qt_cleanup (void *closure)
 cairo_surface_t *
 _cairo_boilerplate_qt_create_surface (const char		 *name,
 				      cairo_content_t		  content,
-				      int			  width,
-				      int			  height,
-				      int			  max_width,
-				      int			  max_height,
+				      double			  width,
+				      double			  height,
+				      double			  max_width,
+				      double			  max_height,
 				      cairo_boilerplate_mode_t	  mode,
 				      int			  id,
 				      void			**closure)
