+++ boilerplate/cairo-boilerplate-script.c
@@ -33,17 +33,17 @@ cairo_user_data_key_t script_closure_key;
 
 typedef struct _script_target_closure {
     char		*filename;
-    int			 width;
-    int			 height;
+    double		 width;
+    double		 height;
 } script_target_closure_t;
 
 cairo_surface_t *
 _cairo_boilerplate_script_create_surface (const char		 *name,
 					  cairo_content_t	  content,
-					  int			  width,
-					  int			  height,
-					  int			  max_width,
-					  int			  max_height,
+					  double			  width,
+					  double			  height,
+					  double			  max_width,
+					  double			  max_height,
 					  cairo_boilerplate_mode_t	  mode,
 					  int                        id,
 					  void			**closure)
