+++ boilerplate/cairo-boilerplate-win32-printing.c
@@ -164,10 +164,10 @@ create_printer_dc (win32_target_closure_t *ptc)
 cairo_surface_t *
 _cairo_boilerplate_win32_printing_create_surface (const char              *name,
 						  cairo_content_t          content,
-						  int                      width,
-						  int                      height,
-						  int                      max_width,
-						  int                      max_height,
+						  double                      width,
+						  double                      height,
+						  double                      max_width,
+						  double                      max_height,
 						  cairo_boilerplate_mode_t mode,
 						  int                      id,
 						  void                   **closure)
