+++ boilerplate/cairo-boilerplate.h
@@ -116,10 +116,10 @@ typedef enum {
 typedef cairo_surface_t *
 (*cairo_boilerplate_create_surface_t) (const char		 *name,
 				       cairo_content_t		  content,
-				       int			  width,
-				       int			  height,
-				       int			  max_width,
-				       int			  max_height,
+				       double			  width,
+				       double			  height,
+				       double			  max_width,
+				       double			  max_height,
 				       cairo_boilerplate_mode_t	  mode,
 				       int                        id,
 				       void			**closure);
